move types to own file
